Fix use of possibly uninitialised variable $LD_LIBRARY_PATH.
With `set -u`, an attempt to use an uninitialised variable in a script
will cause bash to exit with an error. LD_LIBRARY_PATH may not be
initialised, in which case the `run_tests.sh` script fails.

Bash supports the expansion ${var:-alternative} in which $var is used if
set, and the string `alternative` is used if $var is not set. Modify the
run_tests.sh script to use ${LD_LIBRARY_PATH:-} to expand to nothing if
the variable was unset, which in turn suppresses the error relating to
the use of an uninitialised variable.