Set dynamic library paths on OS X.
When System Integrity Protection is enabled, OS X ignores the
DYLD_LIBRARY_PATH environment variable. The clang.cindex.Config class
permits setting the path to libclang from within the Python script.
Similarly, the tests for generated bindings expose a Config that permits
the same setting.

This commit detects OS X and sets the paths appropriately for a default
installation (i.e. using the system compiler). It also sets the path to
the `output` directory in the test scripts, relative to the script path.