OS X fixes (#37)
* Fix use of possibly uninitialised variable $LD_LIBRARY_PATH.

With `set -u`, an attempt to use an uninitialised variable in a script
will cause bash to exit with an error. LD_LIBRARY_PATH may not be
initialised, in which case the `run_tests.sh` script fails.

Bash supports the expansion ${var:-alternative} in which $var is used if
set, and the string `alternative` is used if $var is not set. Modify the
run_tests.sh script to use ${LD_LIBRARY_PATH:-} to expand to nothing if
the variable was unset, which in turn suppresses the error relating to
the use of an uninitialised variable.

* Use clang++-3.8 if available and fall back on clang++ otherwise.

Try to use clang++-3.8 if available in the PATH, otherwise fall back on
clang++ without a version.

* Set dynamic library paths on OS X.

When System Integrity Protection is enabled, OS X ignores the
DYLD_LIBRARY_PATH environment variable. The clang.cindex.Config class
permits setting the path to libclang from within the Python script.
Similarly, the tests for generated bindings expose a Config that permits
the same setting.

This commit detects OS X and sets the paths appropriately for a default
installation (i.e. using the system compiler). It also sets the path to
the `output` directory in the test scripts, relative to the script path.