Create Python script to generate bindings (#52)
Add FFIG.py which has much of the functionality of generate.py but must
be passed arguments to tell it which files to use and generate.

Move clang_cpp_code_model out of externals. I've tried upstreaming this
module into clang but no one seems keen to review it. Life is simpler if
we can import cppmodel easily into FFIG.

Add tests for cppmodel as we now own it.

Update _build.py to call FFIG.py with appropriate arguments.