Generators (#129)
* Support for custom generators.

Some binding types may require custom steps and multiple files to
produce idiomatic output. For example, a Python module needs a directory
and an __init__.py file, and Go packages need a particular directory
structure.

Previously, FFIG would render a single file from a template. Now, it
checks to see whether a given binding has a custom generator function
registered. If so, this custom function is called. If there is no
custom function registered, FFIG falls back on the default, which
replicates the original behaviour of generating a single file in the
base output directory.

Generator plugins may be written in two forms. In either form, they must
expose a function called 'setup_plugin' which accepts a generator
context and registers one or more generator functions. An example might
be:

    def setup_plugin(context):
        context.register(generator_func, ['lang.tmpl'])

where generator_func is a function of the following form (see
generators.default_generator for details of the argument types).

    f(binding, api_classes, env, args, output_dir)

Generator Modules:
    A single .py file in the generators/ directory.
    Suitable for simple generators.

Generator Packages:
    A subdirectory of generators/ with an __init__.py.
    Suitable for complex or compound generators which might use
    submodules to generate different variants of the same language
    (Python2, Python3), (C++98, C++11, C++17), etc.

* Add generator package for Python.

For now, forward to the default generator.

This serves as an example of a package-based generator plugin, rather
than a module-based one.