Codechecks (#137)
* N135: Add codechecks script

Closes #135.

This script runs pep8 over the Python files in the codebase and reports
on the results. If there are any violations, it exits with a non-zero
status so that it can be used in CI.

With the --reformat option, it applies autopep8 first (but still runs
the checks).

It's designed to allow multiple sets of checks and formatting, so each
one will run, even if there were errors in an earlier step. The end
result is only successful if all checks were successful.

Note that the checks won't run if the reformatting failed; this is so
the user has a better chance to see the error from the reformatting
tool.

* N135: Apply autopep8 via ./scripts/codechecks.py --reformat

* N135: Add codechecks to Travis-CI.

* pip install pep8 in Travis-CI