Basic Go bindings. (#145)
* Basic Go bindings.

This commit adds support for simple types in the Go bindings.
It's possible to create an object and call simple functions with simple
return types.

There is no attempt to handle memory management / cleanup, so this can
be assumed to leak resources as implemented.

Go tests are disabled on Windows - the CGo layer doesn't work on AppVeyor. Without a Windows system to debug, I can't make progress on this.