Issue #172: Improve error reporting from Go bindings.
Use `stat` to check for existence of the DSO before trying to `dlopen`
it. This should allow earlier detection of errors involving missing DSO
files or incorrect paths.

Also add some comments about the error codes returned by the `init`
function in the CGo layer, and how to interpret it.