N228: Refactor jinja2 macros. (#229)
* N228: Refactor jinja2 macros.

The common pattern of creating a comma-separated list with optional
leading and trailing commas is factored out for re-use across ffig
templates.

This is achieved by the `call` macro style in jinja2, which essentially
provides a callback whose body is inserted into another macro. See the
example below, or the documentation at
http://jinja.pocoo.org/docs/2.9/templates/#call for more details.

For an outer macro defined as:

    {% macro f1(x) %}
    Outer macro: Hello.
    Inner macro: {{caller(x)}}
    {% endmacro %}

A callback can be provided as follows:

    {% call(y) f1("Hello") %}
    {{y}} again.
    {% endcall %}

This will result in the output:

    Outer macro: Hello.
    Inner macro: Hello again.

This commit closes #228.