N266: Support for generator aliases.
Allow the template used with the default generator to be aliased, e.g.
so that the 'ruby' generator can be used, which maps to 'rb.tmpl'
bindings from the default generator.

There is no point supporting generators other than the default generator
for this mechanism; these can register whichever name they like,
already.