N257 pre push hook (#271)
Add a pre-push hook script that runs the codechecks and blocks a push if the checks fail. This avoids a failing CI cycle.

There's also a script that installs the Git hooks, since they cannot be activated by default. This makes it easier for a developer to bootstrap their environment.