N302: Use lowercase names for Python modules.
To comply with the guidance in
https://www.python.org/dev/peps/pep-0008/#package-and-module-names
this commit converts the module name to lowercase when generating Python
bindings.

This is done in two places: in the CMake definition of the generated
output, and in the generator for Python code.

The tests have been updated to use the lowercase module names.

Closes #302.