N258: Support for noexcept member functions in C++ bindings.
Update the C++ template to generate calls to the `_noexcept` C API
functions when an exposed function is marked `noexcept`.

Add tests to verify that the Tree member functions are noexcept in the
bindings. At the moment, we generate the Tree bindings with the
`NOEXCEPT` flag, which forces exported functions to be `noexcept(true)`.

Ultimately, we'll use libclang to determine whether the function is
`noexcept`.

Closes #258.