N350: Fix Travis macOS build. (#351)
It appears the default in homebrew is for `python` to now refer to python 3, so instead of `brew install python3` we have to `brew update python`.

We then need to install python 2, which FFIG uses to test both Py2 and Py3 functionality. This is done with `brew install python@2`.

The installation of python 2 doesn't add pip2 to the PATH, so the final piece of the puzzle is to use the absolute path to pip2.